backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
pillow==10.4.0
aiofiles==24.1.0
httpx==0.27.2
python-telegram-bot==21.6
aiosmtplib==3.0.1
email-validator==2.2.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
geopy==2.4.1
celery==5.4.0
redis==5.1.1
